feat(errors): create openapi error handling module with structured error classes
